+++ examples/event/event_api_and_broker.py
@@ -5,6 +5,7 @@ from examples.event.event_broker import event_to_handler_map
 from lbz.events import EventBroker
 from lbz.exceptions import LambdaFWException
 from lbz.lambdas import LambdaSource
+from lbz.response import Response
 from lbz.type_defs import LambdaContext
 
 
@@ -16,4 +17,4 @@ def handle(event: dict, context: LambdaContext) -> dict | None:
             EventBroker(mapper=event_to_handler_map, event=event, context=context).react()
         return None
     except Exception:  # pylint: disable=broad-except
-        return LambdaFWException().get_response(context.aws_request_id).to_dict()
+        return Response.from_exception(LambdaFWException(), context.aws_request_id).to_dict()