+++ tests/test_response.py
@@ -3,6 +3,7 @@ from base64 import b64encode
 
 import pytest
 
+from lbz.exceptions import LambdaFWException
 from lbz.response import Response
 
 
@@ -85,6 +86,49 @@ class TestResponse:
             (666, False),
         ],
     )
-    def test_response_is_ok(self, code: int, outcome: bool) -> None:
+    def test_response_ok(self, code: int, outcome: bool) -> None:
         response = Response({"message": "xxx"}, headers={"xx": "xx"}, status_code=code)
+
+        assert response.ok == outcome
         assert response.is_ok() == outcome
+
+    def test_from_exception(self) -> None:
+        response = Response.from_exception(LambdaFWException(), "req-id")
+
+        assert response.to_dict() == {
+            "body": '{"message":"Server got itself in trouble","request_id":"req-id"}',
+            "headers": {"Content-Type": "application/json"},
+            "isBase64Encoded": False,
+            "statusCode": 500,
+        }
+
+    def test_from_exception_with_error_code(self) -> None:
+        class RandomException(LambdaFWException):
+            error_code = "RAND001"
+
+        response = Response.from_exception(RandomException(), "req-id")
+
+        assert response.to_dict() == {
+            "body": '{"message":"Server got itself in '
+            'trouble","request_id":"req-id","error_code":"RAND001"}',
+            "headers": {"Content-Type": "application/json"},
+            "isBase64Encoded": False,
+            "statusCode": 500,
+        }
+
+    def test_json_when_body_is_a_string(self) -> None:
+        response = Response(
+            '{"message":"Server got itself in '
+            'trouble","request_id":"req-id","error_code":"RAND001"}'
+        )
+
+        assert response.json() == {
+            "message": "Server got itself in trouble",
+            "request_id": "req-id",
+            "error_code": "RAND001",
+        }
+
+    def test_json_when_body_is_a_dict(self) -> None:
+        response = Response({"some": "data"})
+
+        assert response.json() == {"some": "data"}